constrained: backwards compatibility with some split sources.
